requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Session==0.5.0

# Database
SQLAlchemy==2.0.25
pyodbc==5.0.1
alembic==1.13.1

# Azure
msal==1.26.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.2.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Security
cryptography==41.0.7
PyJWT==2.8.0
python-dotenv==1.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[flask]==1.39.2
